populating ds6 folder s
