+++ src/ds/CMakeLists.txt
@@ -9,7 +9,16 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/ds-devices-common.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds-motion-common.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds-private.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ds-timestamp.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ds-options.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/rs_advanced_mode.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/presets.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/advanced_mode.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds-devices-common.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds-motion-common.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds-private.h"
+        "${CMAKE_CURRENT_LIST_DIR}/ds-timestamp.h"
+        "${CMAKE_CURRENT_LIST_DIR}/ds-options.h"
+        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/json_loader.hpp"
+        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/presets.h"
 )
