+++ src/ds/ds5/CMakeLists.txt
@@ -3,7 +3,6 @@
 target_sources(${LRS_TARGET}
     PRIVATE
         "${CMAKE_CURRENT_LIST_DIR}/ds5-options.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/ds5-timestamp.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-private.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-motion.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-nonmonochrome.cpp"
@@ -12,23 +11,17 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/ds5-active.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-factory.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-fw-update-device.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/rs_advanced_mode.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/presets.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/advanced_mode.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-auto-calibration.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-thermal-monitor.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-factory.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-device.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-options.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-private.h"
-        "${CMAKE_CURRENT_LIST_DIR}/ds5-timestamp.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-active.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-motion.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-nonmonochrome.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-color.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-fw-update-device.h"
-        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/json_loader.hpp"
-        "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/presets.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-auto-calibration.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds5-thermal-monitor.h"
 )
