+++ src/ds/ds5/ds5-color.cpp
@@ -4,7 +4,7 @@
 #include <cstddef>
 #include "metadata.h"
 
-#include "ds5-timestamp.h"
+#include "ds/ds-timestamp.h"
 #include "ds5-thermal-monitor.h"
 #include "proc/color-formats-converter.h"
 #include "ds5-color.h"
@@ -61,8 +61,8 @@ namespace librealsense
         {
             // means color end point in part of a separate color sensor (e.g. D435)
             color_devs_info = color_devs_info_mi3;
-            std::unique_ptr<frame_timestamp_reader> ds5_timestamp_reader_backup(new ds5_timestamp_reader(backend.create_time_service()));
-            std::unique_ptr<frame_timestamp_reader> ds5_timestamp_reader_metadata(new ds5_timestamp_reader_from_metadata(std::move(ds5_timestamp_reader_backup)));
+            std::unique_ptr<frame_timestamp_reader> ds5_timestamp_reader_backup(new ds_timestamp_reader(backend.create_time_service()));
+            std::unique_ptr<frame_timestamp_reader> ds5_timestamp_reader_metadata(new ds_timestamp_reader_from_metadata(std::move(ds5_timestamp_reader_backup)));
 
             auto enable_global_time_option = std::shared_ptr<global_time_option>(new global_time_option());
             auto raw_color_ep = std::make_shared<uvc_sensor>("Raw RGB Camera",
