+++ src/ds/ds6/CMakeLists.txt
@@ -7,10 +7,10 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/ds6-motion.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds6-active.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ds6-private.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ds6-factory.cpp"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-options.cpp"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-timestamp.cpp"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-nonmonochrome.cpp"
-#       "${CMAKE_CURRENT_LIST_DIR}/ds5-factory.cpp"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-fw-update-device.cpp"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds-devices-common.cpp"
 #       "${CMAKE_CURRENT_LIST_DIR}/advanced_mode/rs_advanced_mode.cpp"
@@ -23,7 +23,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/ds6-motion.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds6-active.h"
         "${CMAKE_CURRENT_LIST_DIR}/ds6-private.h"
-#       "${CMAKE_CURRENT_LIST_DIR}/ds5-factory.h"
+        "${CMAKE_CURRENT_LIST_DIR}/ds6-factory.h"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-options.h"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-timestamp.h"
 #       "${CMAKE_CURRENT_LIST_DIR}/ds5-nonmonochrome.h"
